+++ frontend/modules/stores/controllers/AdminCategoriesController.php
@@ -38,6 +38,11 @@ class AdminCategoriesController extends Controller
    */
   public function actionIndex()
   {
+    if (Yii::$app->user->isGuest ||  !Yii::$app->user->can('CategoriesView')) {
+      throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+      return false;
+    }
+
     $searchModel = new CategoriesStoresSearch();
     $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -59,6 +64,10 @@ class AdminCategoriesController extends Controller
    */
   public function actionCreate()
   {
+    if (Yii::$app->user->isGuest ||  !Yii::$app->user->can('ReviewsCreate')) {
+      throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+      return false;
+    }
     $model = new CategoriesStores();
 
     if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -78,6 +87,10 @@ class AdminCategoriesController extends Controller
    */
   public function actionUpdate($id)
   {
+    if (Yii::$app->user->isGuest ||  !Yii::$app->user->can('CategoriesEdit')) {
+      throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+      return false;
+    }
     $model = $this->findModel($id);
 
     if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -97,6 +110,11 @@ class AdminCategoriesController extends Controller
    */
   public function actionDelete($id)
   {
+    if (Yii::$app->user->isGuest ||  !Yii::$app->user->can('ReviewsDelete')) {
+      throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+      return false;
+    }
+
     $this->findModel($id)->delete();
 
     return $this->redirect(['index']);
